:art: Moved Profile.class to Models
